+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.3.0+1
+
+* Support the latest `coverage` `0.7.0` series releases.
+
 ### 0.3.0
 
 * `serviceName` was removed from `CommandLineClient`.